Move to Breeze.
Look, I'll be honest here.  The switch from Plasma 5 to Plasma 6 has
broken a lot of what I "borrowed" from Sweet and the Garuda
customisations.

I feel that it will be simpler to just use Breeze.  Those are the
blessed themes, so when Plasma 7 lands, I can just re-apply the hacks
and have everything working.